add config to do dry-run of a presentation

+++ tutorial/scripts/02_script.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
-VERSION="0.0.29" #version of the application
+VERSION="0.0.32" #version of the application
 NAMESPACE=$(cat tutorial/scripts/.namespace) #name of your OpenShift namespace
 REBUILD=true #whether or not the application and image need to be rebuild
+CONFIGURE=true #whether or not the config files should already be ajusted
 DOCKER_IMAGE="quay.io/appdev_playground/knative_demo:monolith-uberjar-$VERSION"
 ROUTE_HOST="monolith-$NAMESPACE.apps.ocp4-bm.redhat.arrowlabs.be"
 
@@ -21,28 +22,31 @@ then
   mv application/configuration/monolith-configuration/src/main/resources/application_backup.properties application/configuration/monolith-configuration/src/main/resources/application.properties
 fi
 
-#create deployment
-config="$(cat tutorial/openshift_definitions/02/deployment_config.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-config="$(echo "${config//<DOCKER_IMAGE>/$DOCKER_IMAGE}")"
-echo "$config" > tutorial/openshift_definitions/02/temp_deployment_config.yaml
-oc apply -f tutorial/openshift_definitions/02/temp_deployment_config.yaml
-rm tutorial/openshift_definitions/02/temp_deployment_config.yaml
+if $CONFIGURE
+then
+  #create deployment
+  config="$(cat tutorial/openshift_definitions/02/deployment_config.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  config="$(echo "${config//<DOCKER_IMAGE>/$DOCKER_IMAGE}")"
+  echo "$config" > tutorial/openshift_definitions/02/temp_deployment_config.yaml
+  oc apply -f tutorial/openshift_definitions/02/temp_deployment_config.yaml
+  rm tutorial/openshift_definitions/02/temp_deployment_config.yaml
 
-#create service
-config="$(cat tutorial/openshift_definitions/02/8080_service_config.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-echo "$config" > tutorial/openshift_definitions/02/temp_8080_service_config.yaml
-oc apply -f tutorial/openshift_definitions/02/temp_8080_service_config.yaml
-rm tutorial/openshift_definitions/02/temp_8080_service_config.yaml
+  #create service
+  config="$(cat tutorial/openshift_definitions/02/8080_service_config.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  echo "$config" > tutorial/openshift_definitions/02/temp_8080_service_config.yaml
+  oc apply -f tutorial/openshift_definitions/02/temp_8080_service_config.yaml
+  rm tutorial/openshift_definitions/02/temp_8080_service_config.yaml
 
-#create route
-config="$(cat tutorial/openshift_definitions/02/8080_route_config.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-config="$(echo "${config//<HOST>/$ROUTE_HOST}")"
-echo "$config" > tutorial/openshift_definitions/02/temp_8080_route_config.yaml
-oc apply -f tutorial/openshift_definitions/02/temp_8080_route_config.yaml
-rm tutorial/openshift_definitions/02/temp_8080_route_config.yaml
+  #create route
+  config="$(cat tutorial/openshift_definitions/02/8080_route_config.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  config="$(echo "${config//<HOST>/$ROUTE_HOST}")"
+  echo "$config" > tutorial/openshift_definitions/02/temp_8080_route_config.yaml
+  oc apply -f tutorial/openshift_definitions/02/temp_8080_route_config.yaml
+  rm tutorial/openshift_definitions/02/temp_8080_route_config.yaml
+fi
