+++ tutorial/scripts/04a_script.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
-VERSION="0.0.19" #version of the application
+VERSION="0.0.21" #version of the application
 NAMESPACE=$(cat tutorial/scripts/.namespace) #name of your OpenShift namespace
 REBUILD=true #whether or not the application and image need to be rebuild
+CONFIGURE=true #whether or not the config files should already be ajusted
 APP_DOCKER_IMAGE="quay.io/appdev_playground/knative_demo:microservice-account-uberjar-$VERSION"
 
 if $REBUILD
@@ -20,22 +21,16 @@ then
   mv application/configuration/microservice-account-configuration/src/main/resources/application_backup.properties application/configuration/microservice-account-configuration/src/main/resources/application.properties
 fi
 
-#create microservice account Knative service
-config="$(cat tutorial/openshift_definitions/04a/knative_service_microservice_account.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-config="$(echo "${config//<DOCKER_IMAGE>/$APP_DOCKER_IMAGE}")"
-echo "$config" > tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
-oc apply -f tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
-rm tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
-
-#create monolith_data_changed debezium connector
-config="$(cat tutorial/openshift_definitions/04a/debezium_postgres_connector.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-echo "$config" > tutorial/openshift_definitions/04a/temp_debezium_postgres_connector.yaml
-oc apply -f tutorial/openshift_definitions/04a/temp_debezium_postgres_connector.yaml
-rm tutorial/openshift_definitions/04a/temp_debezium_postgres_connector.yaml
+if $CONFIGURE
+then
+  #create microservice account Knative service
+  config="$(cat tutorial/openshift_definitions/04a/knative_service_microservice_account.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  config="$(echo "${config//<DOCKER_IMAGE>/$APP_DOCKER_IMAGE}")"
+  echo "$config" > tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
+  oc apply -f tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
+  rm tutorial/openshift_definitions/04a/temp_knative_service_microservice_account.yaml
 
 echo ""
 echo ""
@@ -44,3 +39,5 @@ echo "oc exec -it my-cluster-kafka-0 \\ \n\
         -- bin/kafka-console-consumer.sh \\ \n\
         --bootstrap-server my-cluster-kafka-bootstrap.$NAMESPACE.svc.cluster.local:9092 \\ \n\
         --topic <TOPIC>"
+
+fi
