+++ tutorial/scripts/05_script.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
-VERSION="0.1.10" #version of the application
+VERSION="0.1.11" #version of the application
 NAMESPACE=$(cat tutorial/scripts/.namespace) #name of your OpenShift namespace
 REBUILD=true #whether or not the application and image need to be rebuild
+CONFIGURE=true #whether or not the config files should already be ajusted
 DOCKER_IMAGE="quay.io/appdev_playground/knative_demo:microservice-address-uberjar-$VERSION"
 
 if $REBUILD
@@ -20,27 +21,30 @@ then
   mv application/configuration/microservice-address-configuration/src/main/resources/application_backup.properties application/configuration/microservice-address-configuration/src/main/resources/application.properties
 fi
 
-#create microservice account Knative service
-config="$(cat tutorial/openshift_definitions/05/knative_service_microservice_address.yaml )" 
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-config="$(echo "${config//<DOCKER_IMAGE>/$DOCKER_IMAGE}")"
-echo "$config" > tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
-oc apply -f tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
-rm tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
+if $CONFIGURE
+then
+  #create microservice account Knative service
+  config="$(cat tutorial/openshift_definitions/05/knative_service_microservice_address.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  config="$(echo "${config//<DOCKER_IMAGE>/$DOCKER_IMAGE}")"
+  echo "$config" > tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
+  oc apply -f tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
+  rm tutorial/openshift_definitions/05/temp_knative_service_microservice_address.yaml
 
-#enable kafka address data changed channel
-config="$(cat tutorial/openshift_definitions/05/kafka_address_data_changed_channel.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-echo "$config" > tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
-oc apply -f tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
-rm tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
+  #enable kafka address data changed channel
+  config="$(cat tutorial/openshift_definitions/05/kafka_address_data_changed_channel.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  echo "$config" > tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
+  oc apply -f tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
+  rm tutorial/openshift_definitions/05/temp_kafka_address_data_changed_channel.yaml
 
-#add microservice account subscriber on address data changed channel
-config="$(cat tutorial/openshift_definitions/05/microservice_account_on_address_data_changed_subscription.yaml )"
-config="$(echo "${config//<VERSION>/$VERSION}")"
-config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
-echo "$config" > tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
-oc apply -f tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
-rm tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
+  #add microservice account subscriber on address data changed channel
+  config="$(cat tutorial/openshift_definitions/05/microservice_account_on_address_data_changed_subscription.yaml )"
+  config="$(echo "${config//<VERSION>/$VERSION}")"
+  config="$(echo "${config//<NAMESPACE>/$NAMESPACE}")"
+  echo "$config" > tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
+  oc apply -f tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
+  rm tutorial/openshift_definitions/05/temp_microservice_account_on_address_data_changed_subscription.yaml
+fi
